BentoYoloCheck/service/requirements.txt
bentoml
ultralytics
pillow
pydantic
pyyaml
